refactor(shorebird_cli): rename process.dart to shorebird_process.dart
