+++ packages/shorebird_cli/lib/src/artifact_manager.dart
@@ -7,7 +7,7 @@ import 'package:path/path.dart' as p;
 import 'package:scoped/scoped.dart';
 import 'package:shorebird_cli/src/cache.dart';
 import 'package:shorebird_cli/src/http_client/http_client.dart';
-import 'package:shorebird_cli/src/process.dart';
+import 'package:shorebird_cli/src/shorebird_process.dart';
 
 /// A reference to a [ArtifactManager] instance.
 final artifactManagerRef = create(ArtifactManager.new);
