+++ packages/shorebird_cli/lib/src/cache.dart
@@ -9,9 +9,9 @@ import 'package:scoped/scoped.dart';
 import 'package:shorebird_cli/src/http_client/http_client.dart';
 import 'package:shorebird_cli/src/logger.dart';
 import 'package:shorebird_cli/src/platform.dart';
-import 'package:shorebird_cli/src/process.dart';
 import 'package:shorebird_cli/src/shorebird_artifacts.dart';
 import 'package:shorebird_cli/src/shorebird_env.dart';
+import 'package:shorebird_cli/src/shorebird_process.dart';
 
 /// {@template cache_update_failure}
 /// Thrown when a cache update fails.
