+++ packages/shorebird_cli/lib/src/commands/run_command.dart
@@ -5,7 +5,7 @@ import 'dart:io';
 import 'package:shorebird_cli/src/command.dart';
 import 'package:shorebird_cli/src/doctor.dart';
 import 'package:shorebird_cli/src/logger.dart';
-import 'package:shorebird_cli/src/process.dart';
+import 'package:shorebird_cli/src/shorebird_process.dart';
 import 'package:shorebird_cli/src/shorebird_validator.dart';
 
 /// {@template run_command}
