+++ packages/shorebird_cli/lib/src/executables/adb.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 
 import 'package:scoped/scoped.dart';
 import 'package:shorebird_cli/src/android_sdk.dart';
-import 'package:shorebird_cli/src/process.dart';
+import 'package:shorebird_cli/src/shorebird_process.dart';
 
 /// A reference to a [Adb] instance.
 final adbRef = create(Adb.new);
