+++ packages/shorebird_cli/lib/src/executables/bundletool.dart
@@ -4,7 +4,7 @@ import 'package:shorebird_cli/src/android_sdk.dart';
 import 'package:shorebird_cli/src/cache.dart';
 import 'package:shorebird_cli/src/executables/java.dart';
 import 'package:shorebird_cli/src/extensions/string.dart';
-import 'package:shorebird_cli/src/process.dart';
+import 'package:shorebird_cli/src/shorebird_process.dart';
 
 /// A reference to a [Bundletool] instance.
 final bundletoolRef = create(Bundletool.new);
