+++ packages/shorebird_cli/lib/src/executables/git.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:scoped/scoped.dart';
-import 'package:shorebird_cli/src/process.dart';
+import 'package:shorebird_cli/src/shorebird_process.dart';
 
 /// A reference to a [Git] instance.
 final gitRef = create(Git.new);
