+++ packages/shorebird_cli/lib/src/executables/idevicesyslog.dart
@@ -6,8 +6,8 @@ import 'package:path/path.dart' as p;
 import 'package:scoped/scoped.dart';
 import 'package:shorebird_cli/src/executables/devicectl/apple_device.dart';
 import 'package:shorebird_cli/src/logger.dart';
-import 'package:shorebird_cli/src/process.dart';
 import 'package:shorebird_cli/src/shorebird_env.dart';
+import 'package:shorebird_cli/src/shorebird_process.dart';
 
 /// A reference to a [IDeviceSysLog] instance.
 final idevicesyslogRef = create(IDeviceSysLog.new);
