+++ packages/shorebird_cli/lib/src/executables/ios_deploy.dart
@@ -7,8 +7,8 @@ import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
 import 'package:scoped/scoped.dart';
 import 'package:shorebird_cli/src/logger.dart';
-import 'package:shorebird_cli/src/process.dart';
 import 'package:shorebird_cli/src/shorebird_env.dart';
+import 'package:shorebird_cli/src/shorebird_process.dart';
 
 /// A reference to a [IOSDeploy] instance.
 final iosDeployRef = create(IOSDeploy.new);
