+++ packages/shorebird_cli/lib/src/shorebird_build_mixin.dart
@@ -5,9 +5,9 @@ import 'package:mason_logger/mason_logger.dart';
 import 'package:shorebird_cli/src/command.dart';
 import 'package:shorebird_cli/src/logger.dart';
 import 'package:shorebird_cli/src/os/operating_system_interface.dart';
-import 'package:shorebird_cli/src/process.dart';
 import 'package:shorebird_cli/src/shorebird_artifacts.dart';
 import 'package:shorebird_cli/src/shorebird_engine_config.dart';
+import 'package:shorebird_cli/src/shorebird_process.dart';
 
 enum Arch {
   arm64,
