+++ packages/shorebird_cli/test/src/fakes.dart
@@ -1,6 +1,6 @@
 import 'package:http/http.dart' as http;
 import 'package:mocktail/mocktail.dart';
-import 'package:shorebird_cli/src/process.dart';
+import 'package:shorebird_cli/src/shorebird_process.dart';
 import 'package:shorebird_code_push_client/shorebird_code_push_client.dart';
 
 class FakeBaseRequest extends Fake implements http.BaseRequest {}
